Change the helpers to non-static class because VB & C# perform differently
