+++ src/NetAnalyzers/CSharp/Microsoft.CodeQuality.Analyzers/QualityGuidelines/CSharpAvoidMultipleEnumerationsAnalyzer.AvoidMultipleEnumerationsHelpers.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the MIT license.  See License.txt in the project root for license information.
+
+using Microsoft.CodeQuality.Analyzers.QualityGuidelines.AvoidMultipleEnumerations;
+
+namespace Microsoft.CodeAnalysis.CSharp.NetAnalyzers.Microsoft.CodeQuality.Analyzers.QualityGuidelines
+{
+    internal partial class CSharpAvoidMultipleEnumerationsAnalyzer
+    {
+        private class CSharpAvoidMultipleEnumerationsHelpers : AvoidMultipleEnumerationsHelpers
+        {
+            public static readonly CSharpAvoidMultipleEnumerationsHelpers Instance = new();
+        }
+    }
+}
