+++ src/NetAnalyzers/VisualBasic/Microsoft.CodeQuality.Analyzers/QualityGuidelines/BasicAvoidMultipleEnumerationsAnalyzer.DataFlowOperationVisitor.vb
@@ -1,11 +1,12 @@
 ﻿' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the MIT license.  See License.txt in the project root for license information.
+
 Imports Microsoft.CodeAnalysis
 Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
 Imports Microsoft.CodeQuality.Analyzers.QualityGuidelines.AvoidMultipleEnumerations
 Imports Microsoft.CodeQuality.Analyzers.QualityGuidelines.AvoidMultipleEnumerations.FlowAnalysis
 
 Namespace Microsoft.CodeQuality.VisualBasic.Analyzers.QualityGuidelines
-    Partial Public NotInheritable Class BasicAvoidMultipleEnumerationsAnalyzer
+    Partial Friend NotInheritable Class BasicAvoidMultipleEnumerationsAnalyzer
         Private Class BasicInvocationCountDataFlowOperationVisitor
             Inherits AvoidMultipleEnumerationsFlowStateDictionaryFlowOperationVisitor
